split covector tests into CLI and integration
