+++ packages/covector/test/cli/command-init.test.ts
@@ -18,7 +18,7 @@ describe("integration test for init command", () => {
           /^\? What is the name of your default branch\? \(main\)$/,
           "pressEnter",
         ],
-      ],
+      ]
     );
 
     expect(stderr).toBe("");
@@ -43,7 +43,7 @@ describe("integration test for init command", () => {
           /\? What is the name of your default branch\? \(main\)$/,
           "pressEnter",
         ],
-      ],
+      ]
     );
 
     expect(stderr).toBe("");
@@ -54,7 +54,7 @@ describe("integration test for init command", () => {
     const config = yield loadFile("./.changes/config.json", fullIntegration);
     expect(config.path).toEqual(".changes/config.json");
     expect(JSON.parse(config.content).gitSiteUrl).toBe(
-      "https://www.github.com/jbolda/covector/",
+      "https://www.github.com/jbolda/covector/"
     );
   }, 25000); // windows takes some time
 });
