+++ packages/covector/test/cli/command-status.test.ts
