+++ packages/covector/test/cli/helpers.ts
