+++ packages/covector/test/integration/preMode.test.ts
@@ -1,4 +1,4 @@
-import { covector } from "../src";
+import { covector } from "../../src";
 import { CovectorVersion } from "@covector/types";
 import { it, captureError } from "@effection/jest";
 import { loadFile } from "@covector/files";
@@ -24,7 +24,7 @@ describe("integration test with preMode `on`", () => {
     "tag": "beta",
     "changes": [${prevChanges.length === 0 ? "" : prevChanges.join(", ")}]
   }
-  `,
+  `
     );
 
   beforeEach(() => {
@@ -53,25 +53,25 @@ describe("integration test with preMode `on`", () => {
 
     const changelogTauriCore = yield loadFile(
       path.join("/tauri/", "CHANGELOG.md"),
-      fullIntegration,
+      fullIntegration
     );
     // has a direct minor from 0.5.2
     expect(changelogTauriCore.content).toBe(
       "# Changelog\n\n" +
         "## \\[0.6.0-beta.0]\n\n" +
-        "- Summary about the changes in tauri\n",
+        "- Summary about the changes in tauri\n"
     );
 
     const changelogTaurijs = yield loadFile(
       path.join("/cli/tauri.js/", "CHANGELOG.md"),
-      fullIntegration,
+      fullIntegration
     );
     // tauri.js through a dep bump
     expect(changelogTaurijs.content).toBe(
       "# Changelog\n\n" +
         "## \\[0.6.3-beta.0]\n\n" +
         "### Dependencies\n\n" +
-        "- Upgraded to `tauri@0.6.0-beta.0`\n",
+        "- Upgraded to `tauri@0.6.0-beta.0`\n"
     );
   });
 
@@ -86,17 +86,17 @@ describe("integration test with preMode `on`", () => {
 
     const changelogTauriCoreOne = yield loadFile(
       path.join("/tauri/", "CHANGELOG.md"),
-      fullIntegration,
+      fullIntegration
     );
     expect(changelogTauriCoreOne.content).toBe(
       "# Changelog\n\n" +
         "## \\[0.6.0-beta.0]\n\n" +
-        "- Summary about the changes in tauri\n",
+        "- Summary about the changes in tauri\n"
     );
 
     const changelogTaurijsOne = yield loadFile(
       path.join("/cli/tauri.js/", "CHANGELOG.md"),
-      fullIntegration,
+      fullIntegration
     );
     // tauri.js does not have a change file directly or through a dep bump
     // so it should remain the same
@@ -104,15 +104,15 @@ describe("integration test with preMode `on`", () => {
       "# Changelog\n\n" +
         "## \\[0.6.3-beta.0]\n\n" +
         "### Dependencies\n\n" +
-        "- Upgraded to `tauri@0.6.0-beta.0`\n",
+        "- Upgraded to `tauri@0.6.0-beta.0`\n"
     );
 
     const preOne = yield loadFile(
       path.join(".changes", "pre.json"),
-      fullIntegration,
+      fullIntegration
     );
     expect(preOne.content).toBe(
-      '{\n  "tag": "beta",\n  "changes": [\n    ".changes/first-change.md",\n    ".changes/second-change.md"\n  ]\n}\n',
+      '{\n  "tag": "beta",\n  "changes": [\n    ".changes/first-change.md",\n    ".changes/second-change.md"\n  ]\n}\n'
     );
 
     // add change file
@@ -123,16 +123,16 @@ describe("integration test with preMode `on`", () => {
 ---
 
 Boop again.
-`,
+`
     );
 
     // double check the write and formatting
     const newChange = yield loadFile(
       path.join(".changes", "third-change.md"),
-      fullIntegration,
+      fullIntegration
     );
     expect(newChange.content).toBe(
-      "---\n" + '"tauri-api": patch\n' + "---\n\n" + "Boop again.\n",
+      "---\n" + '"tauri-api": patch\n' + "---\n\n" + "Boop again.\n"
     );
 
     const covectoredTwo = (yield covector({
@@ -142,7 +142,7 @@ Boop again.
 
     const changelogTauriCoreTwo = yield loadFile(
       path.join("/tauri/", "CHANGELOG.md"),
-      fullIntegration,
+      fullIntegration
     );
     expect(changelogTauriCoreTwo.content).toBe(
       "# Changelog\n\n" +
@@ -151,12 +151,12 @@ Boop again.
         "- Upgraded to `tauri-api@0.5.2-beta.0`\n" +
         "\n" +
         "## \\[0.6.0-beta.0]\n\n" +
-        "- Summary about the changes in tauri\n",
+        "- Summary about the changes in tauri\n"
     );
 
     const changelogTaurijsTwo = yield loadFile(
       path.join("/cli/tauri.js/", "CHANGELOG.md"),
-      fullIntegration,
+      fullIntegration
     );
     // tauri.js does not have a change file directly or through a dep bump
     // so it should remain the same
@@ -168,15 +168,15 @@ Boop again.
         "\n" +
         "## \\[0.6.3-beta.0]\n\n" +
         "### Dependencies\n\n" +
-        "- Upgraded to `tauri@0.6.0-beta.0`\n",
+        "- Upgraded to `tauri@0.6.0-beta.0`\n"
     );
 
     const preTwo = yield loadFile(
       path.join(".changes", "pre.json"),
-      fullIntegration,
+      fullIntegration
     );
     expect(preTwo.content).toBe(
-      '{\n  "tag": "beta",\n  "changes": [\n    ".changes/first-change.md",\n    ".changes/second-change.md",\n    ".changes/third-change.md"\n  ]\n}\n',
+      '{\n  "tag": "beta",\n  "changes": [\n    ".changes/first-change.md",\n    ".changes/second-change.md",\n    ".changes/third-change.md"\n  ]\n}\n'
     );
 
     if (typeof covectoredOne !== "object")
@@ -211,13 +211,13 @@ Boop again.
     }).toMatchSnapshot();
 
     const changelogTauriCore = yield captureError(
-      loadFile(path.join("/tauri/", "CHANGELOG.md"), fullIntegration),
+      loadFile(path.join("/tauri/", "CHANGELOG.md"), fullIntegration)
     );
     expect(changelogTauriCore.effectionTrace[0].state).toEqual("erroring");
     expect(changelogTauriCore.effectionTrace[1].state).toEqual("erroring");
 
     const changelogTaurijs = yield captureError(
-      loadFile(path.join("/cli/tauri.js/", "CHANGELOG.md"), fullIntegration),
+      loadFile(path.join("/cli/tauri.js/", "CHANGELOG.md"), fullIntegration)
     );
     expect(changelogTaurijs.effectionTrace[0].state).toEqual("erroring");
     expect(changelogTaurijs.effectionTrace[1].state).toEqual("erroring");
